move from mirror to steal test file.
